pi frags are highp

+++ demo/Vert/vertdemo.frag
@@ -1,5 +1,6 @@
 #version 320 es
-precision mediump float;
+precision highp float;
+
 in vec4 v_color;
 out vec4 fragColor;
